backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.27
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
email-validator>=2.1.0.post1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
aiohttp>=3.9.1
redis>=5.0.1
aioredis>=2.0.1

# Celery e Redis
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0  # Interface web para monitorar o Celery

# Processamento de vídeo
yt-dlp>=2023.3.4  # Download de vídeos do YouTube
moviepy>=1.0.3  # Manipulação de vídeo e áudio 